+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/providers/JSONFormDefinitionProvider/json-form-definition.context.ts
@@ -1,9 +1,9 @@
-import { JSONFormElementParams } from '@/components/organisms/UIRenderer/elements/JSONForm/JSONForm';
+import { JSONFormElementBaseParams } from '@/components/organisms/UIRenderer/elements/JSONForm/JSONForm';
 import { UIElement } from '@/domains/collection-flow';
 import { createContext } from 'react';
 
 export interface JSONFormDefinitionContext {
-  definition: UIElement<JSONFormElementParams>;
+  definition: UIElement<JSONFormElementBaseParams>;
 }
 
 export const jsonFormDefinitionContext = createContext<JSONFormDefinitionContext>(