+++ packages/ui/src/components/organisms/DynamicForm/components/layouts/FieldLayout/FieldLayout.tsx
@@ -14,7 +14,8 @@ export const FieldLayout = ({
   description,
   uiSchema,
   required,
-}: FieldTemplateProps) => {
+  optionalLabel = ' (optional) ',
+}: FieldTemplateProps & { optionalLabel?: string }) => {
   const { fieldWarnings } = useWarnings(id);
 
   const isLabelEnabled = displayLabel || (uiSchema && uiSchema['ui:label']);
@@ -27,7 +28,7 @@ export const FieldLayout = ({
       {isLabelEnabled ? (
         <Label htmlFor={id}>
           {label}
-          {required ? '' : <span className="opacity-50">{' (optional) '}</span>}
+          {required ? '' : <span className="opacity-50">{optionalLabel}</span>}
         </Label>
       ) : null}
       {children}