+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ballerine/backoffice-v2
 
+## 0.5.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.3.21
+
+## 0.5.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.3.20
+
 ## 0.5.34
 
 ### Patch Changes