+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # kyb-app
 
+## 0.1.31
+
+### Patch Changes
+
+- Translating optional label
+- Updated dependencies
+  - @ballerine/ui@0.3.21
+
+## 0.1.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.3.20
+
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/blocks@0.1.23
+
 ## 0.1.28
 
 ### Patch Changes