+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/headless-example
 
+## 0.1.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.33
+  - @ballerine/workflow-browser-sdk@0.5.32
+
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.32
+  - @ballerine/workflow-browser-sdk@0.5.31
+
 ## 0.1.30
 
 ### Patch Changes